# swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "CareTogether CMS API",
    "description": "API for the CareTogether Community Management System",
    "contact": {
      "name": "CareTogether CMS Team",
      "url": "https://caretogether.io/"
    },
    "license": {
      "name": "Use under AGPLv3",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/{organizationId}/{locationId}/Directory": {
      "get": {
        "tags": [
          "Directory"
        ],
        "operationId": "Directory_ListVisibleFamilies",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CombinedFamilyInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/{organizationId}/{locationId}/Directory/directoryCommand": {
      "post": {
        "tags": [
          "Directory"
        ],
        "operationId": "Directory_SubmitDirectoryCommand",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "command",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectoryCommand"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombinedFamilyInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/{organizationId}/{locationId}/Directory/familyCommand": {
      "post": {
        "tags": [
          "Directory"
        ],
        "operationId": "Directory_SubmitFamilyCommand",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          },
          {
            "name": "familyId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "command",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyCommand"
              }
            }
          },
          "required": true,
          "x-position": 4
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombinedFamilyInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/{organizationId}/{locationId}/Directory/personCommand": {
      "post": {
        "tags": [
          "Directory"
        ],
        "operationId": "Directory_SubmitPersonCommand",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          },
          {
            "name": "familyId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "command",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCommand"
              }
            }
          },
          "required": true,
          "x-position": 4
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombinedFamilyInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/{organizationId}/{locationId}/Directory/noteCommand": {
      "post": {
        "tags": [
          "Directory"
        ],
        "operationId": "Directory_SubmitNoteCommand",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "command",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCommand"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteCommandResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CombinedFamilyInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "family": {
            "$ref": "#/components/schemas/Family"
          },
          "partneringFamilyInfo": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PartneringFamilyInfo"
              }
            ]
          },
          "volunteerFamilyInfo": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/VolunteerFamilyInfo"
              }
            ]
          },
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Note"
            }
          },
          "uploadedDocuments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadedDocumentInfo"
            }
          }
        }
      },
      "Family": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "primaryFamilyContactPersonId": {
            "type": "string",
            "format": "guid"
          },
          "adults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValueTupleOfPersonAndFamilyAdultRelationshipInfo"
            }
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Person"
            }
          },
          "custodialRelationships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustodialRelationship"
            }
          },
          "uploadedDocuments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadedDocumentInfo"
            }
          },
          "deletedDocuments": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "guid"
            }
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            }
          }
        }
      },
...
      "UploadedDocumentInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "userId": {
            "type": "string",
            "format": "guid"
          },
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          },
          "uploadedDocumentId": {
            "type": "string",
            "format": "guid"
          },
          "uploadedFileName": {
            "type": "string"
          }
        }
      },
...
      "PartneringFamilyInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "openReferral": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Referral"
              }
            ]
          },
          "closedReferrals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Referral"
            }
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            }
          }
        }
      },
...
      "VolunteerFamilyInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "completedRequirements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletedRequirementInfo"
            }
          },
          "exemptedRequirements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExemptedRequirementInfo"
            }
          },
          "removedRoles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RemovedRole"
            }
          },
          "missingRequirements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "availableApplications": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "familyRoleApprovals": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/RoleVersionApproval"
              }
            }
          },
          "individualVolunteers": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/VolunteerInfo"
            }
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            }
          }
        }
      },
...
      "Note": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "authorId": {
            "type": "string",
            "format": "guid"
          },
          "timestampUtc": {
            "type": "string",
            "format": "date-time"
          },
          "contents": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/NoteStatus"
          }
        }
      },
...
      "DirectoryCommand": {
        "type": "object",
        "discriminator": {
          "propertyName": "discriminator",
          "mapping": {
            "AddAdultToFamilyCommand": "#/components/schemas/AddAdultToFamilyCommand",
            "AddChildToFamilyCommand": "#/components/schemas/AddChildToFamilyCommand",
            "CreatePartneringFamilyWithNewAdultCommand": "#/components/schemas/CreatePartneringFamilyWithNewAdultCommand",
            "CreateVolunteerFamilyWithNewAdultCommand": "#/components/schemas/CreateVolunteerFamilyWithNewAdultCommand"
          }
        },
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "discriminator"
        ],
        "properties": {
          "discriminator": {
            "type": "string"
          }
        }
      },
      "AddAdultToFamilyCommand": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DirectoryCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "familyId": {
                "type": "string",
                "format": "guid"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "gender": {
                "$ref": "#/components/schemas/Gender"
              },
              "age": {
                "$ref": "#/components/schemas/Age"
              },
              "ethnicity": {
                "type": "string"
              },
              "familyAdultRelationshipInfo": {
                "$ref": "#/components/schemas/FamilyAdultRelationshipInfo"
              },
              "concerns": {
                "type": "string",
                "nullable": true
              },
              "notes": {
                "type": "string",
                "nullable": true
              },
              "address": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Address"
                  }
                ]
              },
              "phoneNumber": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/PhoneNumber"
                  }
                ]
              },
              "emailAddress": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/EmailAddress"
                  }
                ]
              }
            }
          }
        ]
      },
      "AddChildToFamilyCommand": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DirectoryCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "familyId": {
                "type": "string",
                "format": "guid"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "gender": {
                "$ref": "#/components/schemas/Gender"
              },
              "age": {
                "$ref": "#/components/schemas/Age"
              },
              "ethnicity": {
                "type": "string"
              },
              "custodialRelationships": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CustodialRelationship"
                }
              },
              "concerns": {
                "type": "string",
                "nullable": true
              },
              "notes": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "CreatePartneringFamilyWithNewAdultCommand": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DirectoryCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "referralOpenedAtUtc": {
                "type": "string",
                "format": "date-time"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "gender": {
                "$ref": "#/components/schemas/Gender"
              },
              "age": {
                "$ref": "#/components/schemas/Age"
              },
              "ethnicity": {
                "type": "string"
              },
              "familyAdultRelationshipInfo": {
                "$ref": "#/components/schemas/FamilyAdultRelationshipInfo"
              },
              "concerns": {
                "type": "string",
                "nullable": true
              },
              "notes": {
                "type": "string",
                "nullable": true
              },
              "address": {
                "$ref": "#/components/schemas/Address"
              },
              "phoneNumber": {
                "$ref": "#/components/schemas/PhoneNumber"
              },
              "emailAddress": {
                "$ref": "#/components/schemas/EmailAddress"
              }
            }
          }
        ]
      },
      "CreateVolunteerFamilyWithNewAdultCommand": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DirectoryCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "gender": {
                "$ref": "#/components/schemas/Gender"
              },
              "age": {
                "$ref": "#/components/schemas/Age"
              },
              "ethnicity": {
                "type": "string"
              },
              "familyAdultRelationshipInfo": {
                "$ref": "#/components/schemas/FamilyAdultRelationshipInfo"
              },
              "concerns": {
                "type": "string",
                "nullable": true
              },
              "notes": {
                "type": "string",
                "nullable": true
              },
              "address": {
                "$ref": "#/components/schemas/Address"
              },
              "phoneNumber": {
                "$ref": "#/components/schemas/PhoneNumber"
              },
              "emailAddress": {
                "$ref": "#/components/schemas/EmailAddress"
              }
            }
          }
        ]
      },
      "FamilyCommand": {
        "type": "object",
        "discriminator": {
          "propertyName": "discriminator",
          "mapping": {
            "AddAdultToFamily": "#/components/schemas/AddAdultToFamily",
            "AddChildToFamily": "#/components/schemas/AddChildToFamily",
            "AddCustodialRelationship": "#/components/schemas/AddCustodialRelationship",
            "CreateFamily": "#/components/schemas/CreateFamily",
            "DeleteUploadedFamilyDocument": "#/components/schemas/DeleteUploadedFamilyDocument",
            "RemoveCustodialRelationship": "#/components/schemas/RemoveCustodialRelationship",
            "UpdateAdultRelationshipToFamily": "#/components/schemas/UpdateAdultRelationshipToFamily",
            "UpdateCustodialRelationshipType": "#/components/schemas/UpdateCustodialRelationshipType",
            "UploadFamilyDocument": "#/components/schemas/UploadFamilyDocument"
          }
        },
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "discriminator"
        ],
        "properties": {
          "familyId": {
            "type": "string",
            "format": "guid"
          },
          "discriminator": {
            "type": "string"
          }
        }
      },
      "AddAdultToFamily": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FamilyCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "adultPersonId": {
                "type": "string",
                "format": "guid"
              },
              "relationshipToFamily": {
                "$ref": "#/components/schemas/FamilyAdultRelationshipInfo"
              }
            }
          }
        ]
      },
      "AddChildToFamily": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FamilyCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "childPersonId": {
                "type": "string",
                "format": "guid"
              },
              "custodialRelationships": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CustodialRelationship"
                }
              }
            }
          }
        ]
      },
      "AddCustodialRelationship": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FamilyCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "custodialRelationship": {
                "$ref": "#/components/schemas/CustodialRelationship"
              }
            }
          }
        ]
      },
      "CreateFamily": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FamilyCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "primaryFamilyContactPersonId": {
                "type": "string",
                "format": "guid"
              },
              "adults": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValueTupleOfGuidAndFamilyAdultRelationshipInfo"
                }
              },
              "children": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "guid"
                }
              },
              "custodialRelationships": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CustodialRelationship"
                }
              }
            }
          }
        ]
      },
...
      "DeleteUploadedFamilyDocument": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FamilyCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "uploadedDocumentId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
      "RemoveCustodialRelationship": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FamilyCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "childPersonId": {
                "type": "string",
                "format": "guid"
              },
              "adultPersonId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
      "UpdateAdultRelationshipToFamily": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FamilyCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "adultPersonId": {
                "type": "string",
                "format": "guid"
              },
              "relationshipToFamily": {
                "$ref": "#/components/schemas/FamilyAdultRelationshipInfo"
              }
            }
          }
        ]
      },
      "UpdateCustodialRelationshipType": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FamilyCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "childPersonId": {
                "type": "string",
                "format": "guid"
              },
              "adultPersonId": {
                "type": "string",
                "format": "guid"
              },
              "type": {
                "$ref": "#/components/schemas/CustodialRelationshipType"
              }
            }
          }
        ]
      },
      "UploadFamilyDocument": {
        "allOf": [
          {
            "$ref": "#/components/schemas/FamilyCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "uploadedDocumentId": {
                "type": "string",
                "format": "guid"
              },
              "uploadedFileName": {
                "type": "string"
              }
            }
          }
        ]
      },
      "PersonCommand": {
        "type": "object",
        "discriminator": {
          "propertyName": "discriminator",
          "mapping": {
            "AddPersonAddress": "#/components/schemas/AddPersonAddress",
            "AddPersonEmailAddress": "#/components/schemas/AddPersonEmailAddress",
            "AddPersonPhoneNumber": "#/components/schemas/AddPersonPhoneNumber",
            "CreatePerson": "#/components/schemas/CreatePerson",
            "UndoCreatePerson": "#/components/schemas/UndoCreatePerson",
            "UpdatePersonAddress": "#/components/schemas/UpdatePersonAddress",
            "UpdatePersonAge": "#/components/schemas/UpdatePersonAge",
            "UpdatePersonConcerns": "#/components/schemas/UpdatePersonConcerns",
            "UpdatePersonEmailAddress": "#/components/schemas/UpdatePersonEmailAddress",
            "UpdatePersonEthnicity": "#/components/schemas/UpdatePersonEthnicity",
            "UpdatePersonGender": "#/components/schemas/UpdatePersonGender",
            "UpdatePersonName": "#/components/schemas/UpdatePersonName",
            "UpdatePersonNotes": "#/components/schemas/UpdatePersonNotes",
            "UpdatePersonPhoneNumber": "#/components/schemas/UpdatePersonPhoneNumber",
            "UpdatePersonUserLink": "#/components/schemas/UpdatePersonUserLink"
          }
        },
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "discriminator"
        ],
        "properties": {
          "personId": {
            "type": "string",
            "format": "guid"
          },
          "discriminator": {
            "type": "string"
          }
        }
      },
      "AddPersonAddress": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PersonCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "address": {
                "$ref": "#/components/schemas/Address"
              },
              "isCurrentAddress": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "AddPersonEmailAddress": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PersonCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "emailAddress": {
                "$ref": "#/components/schemas/EmailAddress"
              },
              "isPreferredEmailAddress": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "AddPersonPhoneNumber": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PersonCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "phoneNumber": {
                "$ref": "#/components/schemas/PhoneNumber"
              },
              "isPreferredPhoneNumber": {
                "type": "boolean"
              }
            }
          }
        ]
      },
...
      "UndoCreatePerson": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PersonCommand"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "UpdatePersonAddress": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PersonCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "address": {
                "$ref": "#/components/schemas/Address"
              },
              "isCurrentAddress": {
                "type": "boolean"
              }
            }
          }
        ]
      },
      "UpdatePersonAge": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PersonCommand"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "age": {
                "$ref": "#/components/schemas/Age"
              }
            }
          }
        ]
      },
...
      "NoteCommandResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "family": {
            "$ref": "#/components/schemas/CombinedFamilyInfo"
          },
          "note": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Note"
              }
            ]
          }
        }
      },
      "NoteCommand": {
        "type": "object",
        "discriminator": {
          "propertyName": "discriminator",
          "mapping": {
            "ApproveNote": "#/components/schemas/ApproveNote",
            "CreateDraftNote": "#/components/schemas/CreateDraftNote",
            "DiscardDraftNote": "#/components/schemas/DiscardDraftNote",
            "EditDraftNote": "#/components/schemas/EditDraftNote"
          }
        },
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "discriminator"
        ],
        "properties": {
          "familyId": {
            "type": "string",
            "format": "guid"
          },
          "noteId": {
            "type": "string",
            "format": "guid"
          },
          "discriminator": {
            "type": "string"
          }
        }
      },
...
    }
  }
}